requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
discord.py==2.3.2
dnspython==2.6.1
Flask==3.0.3
frozenlist==1.4.1
humanreadable==0.4.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mbstrdecoder==1.1.3
multidict==6.0.5
pycparser==2.22
pymongo==4.7.3
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
typepy==1.3.2
urllib3==2.2.1
Werkzeug==3.0.3
yarl==1.9.4
zipp==3.19.0